Reworked the media embed error; completely removed Classy version.
Now only inverts the icon on dark colour scheme, which is now displayed
as a ::before generated content element to facilitate this. This change
has been made because there isn't always a child element to reverse the
invert filter so text can be unreadable on dark schemes in those cases.