modules/omnipedia: WikiNodeRoute counts node edit form viewing a node:
'entity.node.edit_form'

This is to fix an issue with the <main-page> element incorrectly setting
an error if your current date based on the last node viewed was
different from the node being edited, which could happen if you went
straight to the node edit form from another admin page, e.g. the content
overview.