Updated drupal/eu_cookie_compliance (1.14.0 => 1.19.0):
This adds a couple of patches required to fix the withdraw button never
being shown:
https://www.drupal.org/project/eu_cookie_compliance/issues/3257088

Exported eu_cookie_compliance.settings.yml with new keys.

modules/omnipedia:
Adds a workaround to prevent the module from forcing the pop-up to
display on admin pages when withdraw consent is enabled even if admin
pages are set to be excluded.

themes/omnipedia:
Removed the categories workaround; now in stable release.