modules/ambientimpact: Added scroll blocker component:
Provides an API to temporarily prevent scrolling the viewport or a
scrollable element when one or more blocking elements are provided.
Layout shifting due to removal of scrollbars on platforms where they
take up space is prevented using the scrollbar gutter component. The
primary use-case for this is with overlays and other modal elements.