Initial pass at migrating to Yarn:
Note that package.json is a bit of a mess and has many dependencies from
module and theme workspaces. These are there to allow the workspaces to
access them, as otherwise Yarn will block access. If possible, these
should be removed from the top-level package.json and inherited somehow.

Also adds the yarn.BUILD plug-in.