drupal/omnipedia_content: Fix Markdown attachments sometimes missing:
MarkdownAttachmentsFilter: Removed conditional logic:

This fixes instances where aggregated assets could be created missing
the references JavaScript if it was created on a wiki page that didn't
have references, i.e. it didn't vary correctly since we never
implemented a cache context for that. We can reintroduce the conditional
logic at a later date if we can correctly vary the output as mentioned
above.

- Upgrading drupal/omnipedia_content (6.x-dev 9f78f75 => 6.x-dev 9f18dde)