Disabling automatic omnipedia_wiki_node_cdn enqueuing for now:
Reducing the number per invocation and increasing the time between
warming has resulted in a queue that only grows over time because of
how basic the current implementation is, in that it'll re-enqueue all
of them before the queue has been completed. This queue constantly
running is probably reducing performance by keeping the back-end busy
responding to those requests in addition to real humans, so we'll have
to revisit this when we can implement this in a more intelligent way
to not re-enqueue what's not needed.

See issue #3