Disabling the omnipedia_wiki_node_cdn warmer:
This doesn't seem to be working lately in that pages still seem to need
to be rendered and thus take a certain amount of time that's not
attributable to slow network or other factors. Once visited, they're
fast to load thereafter, indicating either the warmer is not resulting
in caching, or it is but it's a different context than a "real" visit
and so not a cache hit.